probably best if the UItime type in the zebra library matches the UI date type, and use int for the same reasons
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@8288 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ src/lib/zl_param.h
@@ -31,8 +31,8 @@ typedef int bool;		/* Boolean variable	*/
 
 struct date_st
 {
-	long	ds_yymmdd;	/* Day portion	*/
-	long	ds_hhmmss;	/* time portion */
+  int	ds_yymmdd;	/* Day portion	*/
+  int	ds_hhmmss;	/* time portion */
 };
 typedef struct date_st date;		/* Date in date/time format */
